requirements.txt
fastapi>=0.110
uvicorn>=0.29
elasticsearch>=8.13
requests>=2.31
pyyaml>=6.0
pytest>=8.0
